expose ContractClass in no_std
