+++ crates/cairo-lang-starknet/src/contract_class/compile.rs
@@ -10,17 +10,13 @@ use cairo_lang_diagnostics::ToOption;
 use cairo_lang_filesystem::ids::CrateId;
 use cairo_lang_lowering::db::LoweringGroup;
 use cairo_lang_lowering::ids::ConcreteFunctionWithBodyId;
-use cairo_lang_sierra as sierra;
 use cairo_lang_sierra_generator::canonical_id_replacer::CanonicalReplacer;
 use cairo_lang_sierra_generator::db::SierraGenGroup;
 use cairo_lang_sierra_generator::replace_ids::{replace_sierra_ids_in_program, SierraIdReplacer};
-use cairo_lang_utils::bigint::{deserialize_big_uint, serialize_big_uint, BigUintAsHex};
 use itertools::{chain, Itertools};
-use num_bigint::BigUint;
-use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
-use crate::abi::{AbiBuilder, Contract};
+use crate::abi::AbiBuilder;
 use crate::aliased::Aliased;
 use crate::allowed_libfuncs::{
     validate_compatible_sierra_version, AllowedLibfuncsError, ListSelector,
@@ -34,9 +30,13 @@ use crate::felt252_serde::{sierra_from_felt252s, sierra_to_felt252s, Felt252Serd
 use crate::plugin::consts::{CONSTRUCTOR_MODULE, EXTERNAL_MODULE, L1_HANDLER_MODULE};
 use crate::starknet_plugin_suite;
 
+use super::{ContractClass, ContractEntryPoint, ContractEntryPoints};
+
 #[cfg(test)]
-#[path = "contract_class_test.rs"]
+#[path = "compile_test.rs"]
 mod test;
+
+const DEFAULT_CONTRACT_CLASS_VERSION: &str = "0.1.0";
 
 #[derive(Error, Debug, Eq, PartialEq)]
 pub enum StarknetCompilationError {
@@ -46,58 +46,18 @@ pub enum StarknetCompilationError {
     AllowedLibfuncsError(#[from] AllowedLibfuncsError),
 }
 
-/// Represents a contract in the Starknet network.
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ContractClass {
-    pub sierra_program: Vec<BigUintAsHex>,
-    pub sierra_program_debug_info: Option<cairo_lang_sierra::debug_info::DebugInfo>,
-    pub contract_class_version: String,
-    pub entry_points_by_type: ContractEntryPoints,
-    pub abi: Option<Contract>,
-}
 impl ContractClass {
     /// Extracts Sierra program from the ContractClass and populates it with debug info if
     /// available.
-    pub fn extract_sierra_program(&self) -> Result<sierra::program::Program, Felt252SerdeError> {
+    pub fn extract_sierra_program(
+        &self,
+    ) -> Result<cairo_lang_sierra::program::Program, Felt252SerdeError> {
         let (_, _, mut sierra_program) = sierra_from_felt252s(&self.sierra_program)?;
         if let Some(info) = &self.sierra_program_debug_info {
             info.populate(&mut sierra_program);
         }
         Ok(sierra_program)
     }
-
-    /// Sanity checks the contract class.
-    /// Currently only checks that if ABI exists, its counts match the entry points counts.
-    pub fn sanity_check(&self) {
-        if let Some(abi) = &self.abi {
-            abi.sanity_check(
-                self.entry_points_by_type.external.len(),
-                self.entry_points_by_type.l1_handler.len(),
-                self.entry_points_by_type.constructor.len(),
-            );
-        }
-    }
-}
-
-const DEFAULT_CONTRACT_CLASS_VERSION: &str = "0.1.0";
-
-#[derive(Clone, Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ContractEntryPoints {
-    #[serde(rename = "EXTERNAL")]
-    pub external: Vec<ContractEntryPoint>,
-    #[serde(rename = "L1_HANDLER")]
-    pub l1_handler: Vec<ContractEntryPoint>,
-    #[serde(rename = "CONSTRUCTOR")]
-    pub constructor: Vec<ContractEntryPoint>,
-}
-
-#[derive(Clone, Default, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ContractEntryPoint {
-    /// A field element that encodes the signature of the called function.
-    #[serde(serialize_with = "serialize_big_uint", deserialize_with = "deserialize_big_uint")]
-    pub selector: BigUint,
-    /// The idx of the user function declaration in the sierra program.
-    pub function_idx: usize,
 }
 
 /// Compile the contract given by path.
